pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-seo-platform"
version = "0.1.0"
description = "AI Visibility Intelligence for SaaS Founders"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "typer[all]>=0.12.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.6.0",
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "spacy>=3.8.0",
    "rapidfuzz>=3.10.0",
    "celery[redis]>=5.4.0",
    "rich>=13.9.0",
    "textstat>=0.7.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "ruff", "mypy"]
semrush = ["semrush-api"]
postgres = ["asyncpg", "psycopg2-binary"]

[project.scripts]
aiseo = "aiseo.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/aiseo"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
